requirements.txt
flask>=2.0
pynetdicom>=2.1.0
pydicom>=2.4.2
pillow>=9.0.0
fpdf>=1.7.2
numpy>=1.21
python-magic>=0.4.27
